+++ src/getStepStartStates.test.ts
@@ -64,6 +64,16 @@ import {
   STEP_RM_CONTAINER_REGISTRIES,
   STEP_RM_CONTAINER_REGISTRY_WEBHOOKS,
 } from './steps/resource-manager/container-registry';
+import {
+  STEP_RM_SERVICE_BUS_NAMESPACES,
+  STEP_RM_SERVICE_BUS_SUBSCRIPTIONS,
+  STEP_RM_SERVICE_BUS_QUEUES,
+  STEP_RM_SERVICE_BUS_TOPICS,
+} from './steps/resource-manager/service-bus';
+import {
+  STEP_RM_CDN_PROFILE,
+  STEP_RM_CDN_ENDPOINTS,
+} from './steps/resource-manager/cdn';
 
 describe('getStepStartStates', () => {
   test('all steps represented', () => {
@@ -120,6 +130,12 @@ describe('getStepStartStates', () => {
       [STEP_RM_PRIVATE_DNS_RECORD_SETS]: { disabled: true },
       [STEP_RM_CONTAINER_REGISTRIES]: { disabled: true },
       [STEP_RM_CONTAINER_REGISTRY_WEBHOOKS]: { disabled: true },
+      [STEP_RM_SERVICE_BUS_NAMESPACES]: { disabled: true },
+      [STEP_RM_SERVICE_BUS_QUEUES]: { disabled: true },
+      [STEP_RM_SERVICE_BUS_TOPICS]: { disabled: true },
+      [STEP_RM_SERVICE_BUS_SUBSCRIPTIONS]: { disabled: true },
+      [STEP_RM_CDN_PROFILE]: { disabled: true },
+      [STEP_RM_CDN_ENDPOINTS]: { disabled: true },
     });
   });
 
@@ -172,6 +188,12 @@ describe('getStepStartStates', () => {
       [STEP_RM_PRIVATE_DNS_RECORD_SETS]: { disabled: true },
       [STEP_RM_CONTAINER_REGISTRIES]: { disabled: true },
       [STEP_RM_CONTAINER_REGISTRY_WEBHOOKS]: { disabled: true },
+      [STEP_RM_SERVICE_BUS_NAMESPACES]: { disabled: true },
+      [STEP_RM_SERVICE_BUS_QUEUES]: { disabled: true },
+      [STEP_RM_SERVICE_BUS_TOPICS]: { disabled: true },
+      [STEP_RM_SERVICE_BUS_SUBSCRIPTIONS]: { disabled: true },
+      [STEP_RM_CDN_PROFILE]: { disabled: true },
+      [STEP_RM_CDN_ENDPOINTS]: { disabled: true },
     });
   });
 
@@ -224,6 +246,12 @@ describe('getStepStartStates', () => {
       [STEP_RM_PRIVATE_DNS_RECORD_SETS]: { disabled: false },
       [STEP_RM_CONTAINER_REGISTRIES]: { disabled: false },
       [STEP_RM_CONTAINER_REGISTRY_WEBHOOKS]: { disabled: false },
+      [STEP_RM_SERVICE_BUS_NAMESPACES]: { disabled: false },
+      [STEP_RM_SERVICE_BUS_QUEUES]: { disabled: false },
+      [STEP_RM_SERVICE_BUS_TOPICS]: { disabled: false },
+      [STEP_RM_SERVICE_BUS_SUBSCRIPTIONS]: { disabled: false },
+      [STEP_RM_CDN_PROFILE]: { disabled: false },
+      [STEP_RM_CDN_ENDPOINTS]: { disabled: false },
     });
   });
 });