+++ src/index.ts
@@ -18,6 +18,8 @@ import { apiManagementSteps } from './steps/resource-manager/api-management';
 import { dnsSteps } from './steps/resource-manager/dns';
 import { privateDnsSteps } from './steps/resource-manager/private-dns';
 import { containerRegistrySteps } from './steps/resource-manager/container-registry';
+import { serviceBusSteps } from './steps/resource-manager/service-bus';
+import { cdnSteps } from './steps/resource-manager/cdn';
 
 export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = {
   instanceConfigFields: {
@@ -66,5 +68,7 @@ export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = 
     ...dnsSteps,
     ...privateDnsSteps,
     ...containerRegistrySteps,
+    ...serviceBusSteps,
+    ...cdnSteps,
   ],
 };