+++ src/steps/resource-manager/utils/createResourceGroupResourceRelationship.test.ts
@@ -44,6 +44,46 @@ describe('#createResourceGroupResourceRelationship', () => {
     });
   });
 
+  test.only('should return direct relationship when resourcegroup (lowercase g) exists in jobState', async () => {
+    const resourceGroupId =
+      '/subscriptions/subscription-id/resourceGroups/resource-group-id';
+    const resourceGroupEntity: Entity = {
+      _class: ['Group'],
+      _type: 'azure_resource_group',
+      _key: resourceGroupId,
+    };
+
+    const context = createMockStepExecutionContext<IntegrationConfig>({
+      instanceConfig,
+      entities: [resourceGroupEntity],
+    });
+
+    const resourceEntity: Entity = {
+      _class: ['Service'],
+      _type: 'azure_keyvault_service',
+      _key:
+        resourceGroupId.replace('resourceGroups', 'resourcegroups') +
+        '/providers/Microsoft/KeyVault/vaults/key-vault-id',
+    };
+
+    const result = await createResourceGroupResourceRelationship(
+      context,
+      resourceEntity,
+    );
+
+    expect(result).toEqual({
+      _class: 'HAS',
+      _fromEntityKey:
+        '/subscriptions/subscription-id/resourceGroups/resource-group-id',
+      _key:
+        '/subscriptions/subscription-id/resourceGroups/resource-group-id|has|/subscriptions/subscription-id/resourcegroups/resource-group-id/providers/Microsoft/KeyVault/vaults/key-vault-id',
+      _toEntityKey:
+        '/subscriptions/subscription-id/resourcegroups/resource-group-id/providers/Microsoft/KeyVault/vaults/key-vault-id',
+      _type: 'azure_resource_group_has_keyvault_service',
+      displayName: 'HAS',
+    });
+  });
+
   test('should throw when resourceGroup does not exist in jobState', async () => {
     const resourceGroupId =
       '/subscriptions/subscription-id/resourceGroups/resource-group-id';